+++ Addons/Schumix.MantisBTRssAddon/Rss/MantisBTRss.cs
@@ -297,10 +297,10 @@ namespace Schumix.MantisBTRssAddon
 				{
 					string language = sLManager.GetChannelLocalization(chan, _servername);
 
-					if(asd.ToBoolean(db["ShortUrl"].ToString()))
+					if(db["ShortUrl"].ToBoolean())
 						link = BitlyApi.ShortenUrl(link).ShortUrl;
 					
-					if(asd.ToBoolean(db["Colors"].ToString()))
+					if(db["Colors"].ToBoolean())
 					{
 						if(title.Contains(SchumixBase.Colon.ToString()))
 						{