+++ Addons/Schumix.SvnRssAddon/Commands/Commands.cs
@@ -446,7 +446,7 @@ namespace Schumix.SvnRssAddon.Commands
 						return;
 					}
 					
-					if(sIRCMessage.Info[7].ToLower() != "true" && sIRCMessage.Info[7].ToLower() != "false")
+					if(!sIRCMessage.Info[7].IsBoolean())
 					{
 						sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetWarningText("ValueIsNotTrueOrFalse", sIRCMessage.Channel, sIRCMessage.ServerName));
 						return;
@@ -455,9 +455,9 @@ namespace Schumix.SvnRssAddon.Commands
 					var db = SchumixBase.DManager.QueryFirstRow("SELECT Colors FROM svninfo WHERE LOWER(Name) = '{0}' And ServerName = '{1}'", sUtilities.SqlEscape(sIRCMessage.Info[6].ToLower()), sIRCMessage.ServerName);
 					if(!db.IsNull())
 					{
-						bool enabled = asd.ToBoolean(db["Colors"].ToString());
+						bool enabled = db["Colors"].ToBoolean();
 						
-						if(asd.ToBoolean(sIRCMessage.Info[7].ToLower()) == enabled)
+						if(sIRCMessage.Info[7].ToBoolean() == enabled)
 						{
 							if(enabled)
 							{
@@ -498,7 +498,7 @@ namespace Schumix.SvnRssAddon.Commands
 						return;
 					}
 					
-					if(sIRCMessage.Info[7].ToLower() != "true" && sIRCMessage.Info[7].ToLower() != "false")
+					if(!sIRCMessage.Info[7].IsBoolean())
 					{
 						sSendMessage.SendChatMessage(sIRCMessage, sLManager.GetWarningText("ValueIsNotTrueOrFalse", sIRCMessage.Channel, sIRCMessage.ServerName));
 						return;
@@ -507,9 +507,9 @@ namespace Schumix.SvnRssAddon.Commands
 					var db = SchumixBase.DManager.QueryFirstRow("SELECT ShortUrl FROM svninfo WHERE LOWER(Name) = '{0}' And ServerName = '{1}'", sUtilities.SqlEscape(sIRCMessage.Info[6].ToLower()), sIRCMessage.ServerName);
 					if(!db.IsNull())
 					{
-						bool enabled = asd.ToBoolean(db["ShortUrl"].ToString());
+						bool enabled = db["ShortUrl"].ToBoolean();
 						
-						if(asd.ToBoolean(sIRCMessage.Info[7].ToLower()) == enabled)
+						if(sIRCMessage.Info[7].ToBoolean() == enabled)
 						{
 							if(enabled)
 							{