+++ Core/Schumix.Framework/Utilities.cs
@@ -448,7 +448,7 @@ namespace Schumix.Framework
 			if(sPlatform.IsWindows)
 			{
 				var mos = new ManagementObjectSearcher("root\\CIMV2", "SELECT * FROM Win32_Processor");
-				return (from ManagementObject mo in mos.Get() select (Regex.Replace(mo["Name"].ToString(), @"\s+", SchumixBase.Space.ToString()))).FirstOrDefault();
+				return (from ManagementObject mo in mos.Get() select (Regex.Replace(Convert.ToString(mo["Name"]), @"\s+", SchumixBase.Space.ToString()))).FirstOrDefault();
 			}
 			else if(sPlatform.IsLinux)
 			{